Task8/requirements.txt
absl-py==0.10.0
albumentations==0.4.6
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
future==0.18.2
google-auth==1.20.1
google-auth-oauthlib==0.4.1
grpcio==1.31.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.7.0
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.1
networkx==2.5
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.4.0.42
opencv-python-headless==4.4.0.42
packaging==20.4
pandas==1.1.1
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==0.9.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.2
Shapely==1.7.1
six==1.15.0
tensorboard==2.2.0
tensorboard-plugin-wit==1.7.0
tifffile==2020.8.25
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
